requirements.txt
PyQt6>=6.7.0
pypdf>=4.0.0
Pillow>=10.0.0
pymupdf>=1.24.0
